Minimize the data in the DB
This is done by removing the need for real ID and Type documents as well
as explicit tracking of the parent field. Date is still being tracked
because it is more useful as a field for comparisons (extracting part of
the id and using that for comparison is not well supported).